added generate qr code in the profile menu ... dropdown
