+++ src/view/com/profile/ProfileMenu.tsx
@@ -39,8 +39,12 @@ import {ReportDialog, useReportDialogControl} from '#/components/ReportDialog'
 
 let ProfileMenu = ({
   profile,
+  isQRCodeModalVisible,
+  setIsQRCodeModalVisible,
 }: {
   profile: Shadow<AppBskyActorDefs.ProfileViewDetailed>
+  isQRCodeModalVisible: boolean
+  setIsQRCodeModalVisible: React.Dispatch<React.SetStateAction<boolean>>
 }): React.ReactNode => {
   const {_} = useLingui()
   const {currentAccount, hasSession} = useSession()
@@ -206,6 +210,15 @@ let ProfileMenu = ({
               <Menu.ItemIcon icon={Share} />
             </Menu.Item>
           </Menu.Group>
+
+          <Menu.Item
+            testID="qrcodeButton"
+            label={_(msg`Generate QR Code`)}
+            onPress={() => setIsQRCodeModalVisible(!isQRCodeModalVisible)}>
+            <Menu.ItemText>
+              <Trans>Generate QR Code</Trans>
+            </Menu.ItemText>
+          </Menu.Item>
 
           {hasSession && (
             <>
